cdc-service/requirements.txt
kafka-python==2.0.2
mysql-connector-python==8.1.0
flask==2.3.3
